Rounded off box coordinates

+++ GeometryTunes/GridView.m
@@ -105,7 +105,7 @@
 
 - (CGPoint) getBoxFromCoords:(CGPoint)pos 
 {
-    CGPoint box = CGPointMake(pos.x / [self getBoxWidth], pos.y / [self getBoxWidth]);
+    CGPoint box = CGPointMake((int)pos.x / [self getBoxWidth], (int)pos.y / [self getBoxWidth]);
     if (box.x > numBoxesX || box.y > numBoxesY)
         return CGPointMake(-1, -1);
     return box;
